+++ registry/default/blocks/data-grid/keyboard/index.ts
@@ -2,3 +2,4 @@
 export { DEFAULT_KEYMAP } from "./default-keymap";
 export { matchKeymap, parseBinding, type KeymapEvent, type ParsedBinding } from "./match-keymap";
 export { isPrintableKey } from "./is-printable-key";
+export { isMacPlatform } from "./platform";